+++ tensorcircuit/backends/numpy_backend.py
@@ -392,7 +392,7 @@ class NumpyBackend(numpy_backend.NumPyBackend, ExtendedBackend):  # type: ignore
         jit_compile: Optional[bool] = None,
         **kws: Any
     ) -> Callable[..., Any]:
-        logger.warning("numpy backend has no jit interface, just do nothing")
+        logger.info("numpy backend has no jit interface, just do nothing")
         return f
         # raise NotImplementedError("numpy backend doesn't support jit compiling")